refactor : Move file's directory
- auth 디렉토리를 도메인에 추가합니다.
- common/jwt/dto를 auth 도메인에 포함되도록합니다.
